pyproject.toml
[project]
name = "social-signal"
version = "0.1.0"
description = "Reddit-based research pipeline for studying public discourse about AI's social impact"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "openai>=1.0",
    "langdetect>=1.0",
    "bleach>=6.0",
    "simhash>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
